requirements.txt
Django>=5.1.7
django-crispy-forms>=2.1
django-environ>=0.11.2
whitenoise>=6.7.0
python-dotenv>=1.1.0
google-cloud-aiplatform>=1.71.1
google-genai>=0.5.0
vertexai
cryptography>=44.0.2
bcrypt>=3.2.0
psycopg2>=2.9.0
bootstrap4
crispy_bootstrap4
cloud-sql-python-connector>=1.2.0
SQLAlchemy>=2.0.0
PyMySQL>=1.0.0
pg8000>=1.29.0  # The Python DB-API driver for PostgreSQL
mysqlclient>=2.0.0
psycopg2-binary>=2.9.10
google-cloud-secret-manager>=2.23.2
gunicorn>=21.0.0
django-phonenumber-field>=8.1.0 
phonenumberslite>=9.0.3
phonenumbers>=9.0.3
django-widget-tweaks>=1.5.0
google-generativeai>=0.8.4
google-cloud-texttospeech>=2.26.0
stripe>=12.0.0
#pillow>=11.1.0
django-storages>=1.14.6
olefile>=0.47
chardet>=5.2.0 
reportlab>=4.4.0
decorator>=5.2.1 
